requirements.txt
# --- Core dependencies ---
huggingface-hub>=0.13.0,<1.0.0
tokenizers>=0.13.0,<0.20.0

# --- Audio processing ---
soundfile>=0.12.1
numpy>=1.26

# --- FastAPI stack ---
fastapi==0.116.1
uvicorn[standard]==0.40.0
python-multipart==0.0.20
python-dotenv==1.0.1
fastapi-mail==1.4.1
starlette==0.40.0

# --- Pydantic ---
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydantic-extra-types==2.10.0

# --- HTTP ---
requests==2.32.3
httpx==0.27.2
urllib3==2.2.3
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
click==8.1.8
h11==0.16.0

# --- Email ---
email-validator==2.3.0
cryptography==42.0.5

# --- Google Generative AI ---
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
googleapis-common-protos==1.63.0
grpcio>=1.60.0
grpcio-status>=1.60.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0

# --- Video download ---
yt-dlp==2024.11.4

# --- Development & Testing ---
flake8==7.1.1
pytest==9.0.2
pytest-cov==7.0.0
